Separated category screen to smaller components.
Modifications qui seront validées :
	nouveau fichier : templates/vue/src/views/dashboard/components/Category/CategoryCard.vue
	modifié :         templates/vue/src/views/dashboard/components/Category/CategoryDialog.vue
	nouveau fichier : templates/vue/src/views/dashboard/components/Category/CategoryList.vue
	modifié :         templates/vue/src/views/dashboard/components/Wallet/WalletCard.vue
	modifié :         templates/vue/src/views/dashboard/components/Wallet/WalletDialog.vue
	modifié :         templates/vue/src/views/dashboard/pages/Category.vue
