+++ templates/vue/src/views/dashboard/components/Category/CategoryDialog.vue
@@ -1,74 +1,31 @@
 <template>
 	<v-dialog v-model="show" max-width="500px">
 		<v-card class="elevation-12">
-			<v-toolbar dark :color="category.colour">
+			<v-toolbar dark :color="decoration.modalColor">
 				<v-toolbar-title>
-					{{ decoration.modalTite }}
+					{{ decoration.modalTitle }}
 				</v-toolbar-title>
 			</v-toolbar>
 			<v-card-text>
-			<form ref="form" @submit.prevent="saveCategory">
-				<v-text-field
-					v-model="category.id"
-					name="id"
-					v-if="false"
-					required
-				></v-text-field>
-
-				<v-text-field
-					v-model="category.name"
-					name="categoryname"
-					label="Category Name"
-					type="text"
-					placeholder="Category Name"
-					required
-				></v-text-field>
-
-				<v-select
-					v-model="decoration.categoryType"
-					:items="type"
-					item-text="name"
-					item-value="id"
-					label="Category Type"
-					v-if="!isUpdate"
-				></v-select>
-
-				<v-row>
-					<v-col>
-						<v-color-picker
-							v-model="category.colour"
-							dot-size="7"
-							hide-mode-switch
-							mode="hexa"
-							:show-swatches="decoration.showSwatches"
-							swatches-max-height="70"
-						></v-color-picker>
-					</v-col>
-					<v-col>
-						<v-checkbox
-							v-model="decoration.showSwatches"
-							label="Show Presets"
-						>
-						</v-checkbox>
-					</v-col>
-				</v-row>
-
-				<v-col class="text-right">
-					 <v-btn type="submit" class="mt-4" color="primary" value="log in">
-						 Save
-					 </v-btn>
-				 </v-col>
-			</form>
-		   </v-card-text>
-	   </v-card>
+				<CategoryCard
+					:category-id="categoryId"
+					:is-update="isUpdate"
+					@colorChangeEmit="onColorChange($event)"
+				/>
+			</v-card-text>
+		</v-card>
 	</v-dialog>
 </template>
 
 <script>
-import { customAxios } from '@/axios'
+import CategoryCard from '@/views/dashboard/components/Category/CategoryCard'
 
 export default {
 	name: 'CategoryDialog',
+
+	components: {
+		CategoryCard,
+	},
 
 	props: {
 		value: Boolean,
@@ -76,109 +33,28 @@ export default {
 		categoryId: Number
 	},
 
-	data() {
-		return {
-			category: {
-				id: -1,
-				name: null,
-				type: null,
-				colour: '#4caf50'
-			},
-			type: [
-				{
-					id: '0',
-					name: 'Revenue',
-				},
-				{
-					id: '1',
-					name: 'Expense',
-				},
-			],
-
-			decoration: {
-				needsUpdate: true,
-				showSwatches: false,
-				categoryType: null,
-				modalTite: null,
+	watch: {
+		categoryId: function(newVal){
+			if(newVal == -1) {
+				this.decoration.modalTitle = "New Category";
+			} else {
+				this.decoration.modalTitle = "Update Category";
 			}
 		}
 	},
 
-	created() {
-		this.updateInfo();
-		this.updateModal();
-	},
-
-	beforeUpdate() {
-		// updated hook catches twice, when the dialog is hidden and shown,
-		// only update when the modal is shown
-		if(this.show == true && this.decoration.needsUpdate == true){
-			this.updateInfo();
-			this.decoration.needsUpdate = false;
-			this.updateModal();
-		} else if(this.show == false){
-			this.decoration.needsUpdate = true;
-
-			// reset value to default
-			this.category.id = -1;
-			this.category.name = null;
-			this.category.type = null;
-			this.category.colour = '#4caf50';
+	data() {
+		return {
+			decoration: {
+				modalTitle: 'New Category',
+				modalColor: '#4caf50'
+			},
 		}
 	},
 
 	methods: {
-		updateModal(){
-			if(this.isUpdate){
-				this.decoration.modalTite = "Update Category";
-			} else {
-				this.decoration.modalTite = "New Category";
-			}
-		},
-
-		updateInfo(){
-			if(this.categoryId != -1){
-				let rootComponent = this;
-				let data = {
-					id: this.categoryId,
-				};
-
-				customAxios
-					.post('api/category/detail', data)
-					.then(response => {
-						let returnedData = response.data.category;
-						rootComponent.category.id = returnedData.id;
-						rootComponent.category.name = returnedData.name;
-						rootComponent.category.type = returnedData.type;
-						rootComponent.decoration.categoryType = rootComponent.type[returnedData.type];
-						rootComponent.category.colour = returnedData.color;
-					});
-			}
-		},
-
-		saveCategory(){
-			let data = {
-				id: this.category.id,
-				name: this.category.name,
-				type: this.decoration.categoryType,
-				colour: this.category.colour
-			};
-			console.log(data);
-
-			customAxios
-				.post('api/category/save', data)
-				.then(response => {
-					swal("Success", 'Category saved', 'success');
-					// TODO: update list after success
-					if(isUpdate){
-
-					} else {
-
-					}
-				})
-				.catch(error => {
-					swal("Oops", error.data, 'error');
-				});
+		onColorChange(dataObject){
+			this.decoration.modalColor = dataObject;
 		}
 	},
